CreateFAQ/backend/pyproject.toml
[tool.poetry]
name = "airopsupskill backend"
version = "0.1.0"
description = ""
authors = ["Saurabh Zinjad <zinjadsaurabh1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
openai = "^1.54.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
